data-simulator/backend/requirements.txt
Flask
Flask-Cors
paho-mqtt
SQLAlchemy
psycopg2-binary
numpy
